Agora há uma autenticação verdadeira no user

+++ src/controllers/UserController.ts
@@ -1,10 +1,12 @@
 import type { FastifyRequest, FastifyReply } from 'fastify';
 import { UserRepository } from '../repositories/UserRepository.js';
+import { RefreshTokenRepository } from '../repositories/RefreshTokenRepository.js';
 import type { User } from '../../generated/prisma/client.js';
 import { Password } from '../domain/Password.js';
 import { Email } from '../domain/Email.js';
 import { Text } from '../domain/Text.js';
-import { ValidationError } from '../errors/errors.js';
+import { UnauthorizedError, ValidationError } from '../errors/errors.js';
+import { JWT } from '../domain/JWT.js';
 
 export class UserController {
     signup = async (request: FastifyRequest<{Body: User}>, reply: FastifyReply) => {
@@ -39,30 +41,87 @@ export class UserController {
         const isValid = await Password.compare(data.password, user.password);
         if (!isValid) throw new ValidationError("Login failed");
 
+        const payload = {
+            id: user.id,
+            email: user.email
+        };
+
+        const accessToken = JWT.generateAccessToken(payload);
+        const {refreshToken, jti} = JWT.generateRefreshToken(payload);
+
+        await RefreshTokenRepository.create(jti, user.id);
+
         return reply.status(200).send({
             id: user.id,
-            acess_token: "WIP",
-            refresh_token: "WIP"
+            access_token: accessToken,
+            refresh_token: refreshToken
         });  
     };
 
-    refresh = async (request: FastifyRequest, reply: FastifyReply) => {
-        return reply.status(200).send({
-            status: "to trabalhando nisso ainda"
+    refresh = async (request: FastifyRequest<{Body: {refresh_token: string}}>, reply: FastifyReply) => {
+        const refreshToken = request.body.refresh_token;
+
+        if (!refreshToken) {
+            throw new ValidationError("Refresh token missing");
+        }
+
+        let decoded: any;
+
+        try {
+            console.log(refreshToken);
+            decoded = JWT.verifyRefreshToken(refreshToken);
+        } catch {
+            throw new ValidationError("Invalid refresh token");
+        }
+
+        const exists = await RefreshTokenRepository.exists(decoded.jti);
+
+        if (!exists) {
+            throw new ValidationError("Invalid token");
+        }
+
+        await RefreshTokenRepository.delete(decoded.jti);
+
+        const { refreshToken: newRefreshToken, jti: newJti } = JWT.generateRefreshToken({
+            id: decoded.id,
+            email: decoded.email
+        });
+
+        await RefreshTokenRepository.create(newJti, decoded.id);
+
+        const newAccessToken = JWT.generateAccessToken({
+            id: decoded.id,
+            email: decoded.email
+        });
+
+        return reply.send({
+            access_token: newAccessToken,
+            refresh_token: newRefreshToken
         });
     };
 
-    logout = async (request: FastifyRequest, reply: FastifyReply) => {
-        return reply.status(200).send({
-            status: "to trabalhando nisso ainda"
+    logout = async (request: FastifyRequest<{Body: {refresh_token: string}}>, reply: FastifyReply) => {
+        const { refresh_token } = request.body;
+
+        if (!refresh_token) {
+            throw new ValidationError("Refresh token missing");
+        }
+
+        const decoded = JWT.verifyRefreshToken(refresh_token) as any;
+
+        await RefreshTokenRepository.delete(decoded.jti);
+
+        return reply.send({
+            status: "Logged out"
         });
     };
 
-    getUser = async (request: FastifyRequest<{Params: {id: string}}>, reply: FastifyReply) => {
-        const id = request.params.id;
+    getMe = async (request: FastifyRequest, reply: FastifyReply) => {
+        const id = request.user?.id;
+        if (!id) throw new UnauthorizedError("Token is invalid");
+
         const user = await UserRepository.findById(id);
-
-        if (!user) throw new ValidationError("User not found");
+        if (!user) throw new UnauthorizedError("You cannot see this user");
 
         return reply.status(200).send({
             username: user.username,
@@ -70,8 +129,9 @@ export class UserController {
         });
     };
 
-    updateMe = async (request: FastifyRequest<{Params: {id: string}; Body: User}>, reply: FastifyReply) => {
-        const { id } = request.params;
+    updateMe = async (request: FastifyRequest<{Body: User}>, reply: FastifyReply) => {
+        const id = request.user?.id;
+        if (!id) throw new UnauthorizedError("Token is invalid");
         const data = request.body;
 
         if (Object.keys(data).length === 0) {
@@ -95,8 +155,9 @@ export class UserController {
         });
     };
 
-    deleteMe = async (request: FastifyRequest<{Params: {id: string}}>, reply: FastifyReply) => {
-        const id = request.params.id;
+    deleteMe = async (request: FastifyRequest, reply: FastifyReply) => {
+        const id = request.user?.id;
+        if (!id) throw new UnauthorizedError("Token is invalid");
 
         const user = await UserRepository.delete(id);
